@@ -30,16 +30,16 @@
     ],
     "homepage": "https://github.com/FriendsOfSymfony/FOSElasticaBundle",
     "require": {
-        "php": "^7.4 || ^8.1",
-        "pagerfanta/pagerfanta": "^2.4 || ^3.0 || ^4.0",
-        "psr/log": "^1.0 || ^2.0 || ^3.0",
-        "ruflin/elastica": "^7.1",
-        "symfony/console": "^5.4 || ^6.4 || ^7.1",
-        "symfony/dependency-injection": "^5.4 || ^6.4 || ^7.1",
-        "symfony/event-dispatcher": "^5.4 || ^6.4 || ^7.1",
-        "symfony/framework-bundle": "^5.4 || ^6.4 || ^7.1",
-        "symfony/property-access": "^5.4 || ^6.4 || ^7.1",
-        "symfony/stopwatch": "^5.4 || ^6.4 || ^7.1"
+        "php": "^8.1",
+        "pagerfanta/pagerfanta": "^3.0 || ^4.0",
+        "psr/log": "^3.0",
+        "ruflin/elastica": "^8.0",
+        "symfony/console": "^6.4 || ^7.1",
+        "symfony/dependency-injection": "^6.4 || ^7.1",
+        "symfony/event-dispatcher": "^6.4 || ^7.1",
+        "symfony/framework-bundle": "^6.4 || ^7.1",
+        "symfony/property-access": "^6.4 || ^7.1",
+        "symfony/stopwatch": "^6.4 || ^7.1"
     },
     "require-dev": {
         "doctrine/doctrine-bundle": "^2.1.1",
@@ -47,22 +47,25 @@
         "doctrine/orm": "^2.8 || ^3.2",
         "doctrine/phpcr-odm": "^1.5.3 || ^2.0",
         "ergebnis/composer-normalize": "^2.28",
-        "jackalope/jackalope-doctrine-dbal": "^1.2 || ^2.0",
-        "jms/serializer": "^3.8",
-        "jms/serializer-bundle": "^4.0 || ^5.0",
-        "knplabs/knp-components": "^2.4 || ^3.0 || ^4.0",
-        "php-cs-fixer/shim": "3.8.0",
+        "jackalope/jackalope-doctrine-dbal": "^2.0",
+        "jms/serializer": "^3.32",
+        "jms/serializer-bundle": "^5.0",
+        "knplabs/knp-components": "^4.0 || ^5.0",
+        "php-cs-fixer/shim": "3.60.0",
         "phpstan/extension-installer": "^1.1",
-        "phpstan/phpstan": "^1.4",
-        "phpstan/phpstan-phpunit": "^1.1",
-        "phpstan/phpstan-symfony": "^1.2",
+        "phpstan/phpstan": "^1.12",
+        "phpstan/phpstan-phpunit": "^1.4",
+        "phpstan/phpstan-symfony": "^1.4",
         "phpunit/phpunit": "^9.5",
-        "symfony/expression-language": "^5.4 || ^6.4 || ^7.1",
-        "symfony/messenger": "^5.4 || ^6.4 || ^7.1",
-        "symfony/serializer": "^5.4 || ^6.4 || ^7.1",
-        "symfony/twig-bundle": "^5.4 || ^6.4 || ^7.1",
-        "symfony/web-profiler-bundle": "^5.4 || ^6.4 || ^7.1",
-        "symfony/yaml": "^5.4 || ^6.4 || ^7.1"
+        "symfony/expression-language": "^6.4 || ^7.1",
+        "symfony/messenger": "^6.4 || ^7.1",
+        "symfony/serializer": "^6.4 || ^7.1",
+        "symfony/twig-bundle": "^6.4 || ^7.1",
+        "symfony/web-profiler-bundle": "^6.4 || ^7.1",
+        "symfony/yaml": "^6.4 || ^7.1"
+    },
+    "conflict": {
+        "phpstan/phpstan": "<1.12.0"
     },
     "suggest": {
         "enqueue/elastica-bundle": "For populating Elasticsearch indexes asynchronously and using significanly less resources. Uses Enqueue.",
