@@ -11,6 +11,7 @@
 
 namespace FOS\ElasticaBundle\DependencyInjection;
 
+use Elastic\Elasticsearch\Transport\RequestOptions;
 use FOS\ElasticaBundle\Serializer\Callback as SerializerCallback;
 use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
@@ -319,98 +320,69 @@ class Configuration implements ConfigurationInterface
             ->children()
                 ->arrayNode('clients')
                     ->useAttributeAsKey('id')
+                    ->fixXmlConfig('header')
                     ->prototype('array')
                         ->performNoDeepMerging()
                         // Elastica names its properties with camel case, support both
-                        ->beforeNormalization()
-                        ->ifTrue(function ($v) {
-                            return isset($v['connection_strategy']);
-                        })
-                        ->then(function ($v) {
-                            $v['connectionStrategy'] = $v['connection_strategy'];
-                            unset($v['connection_strategy']);
-
-                            return $v;
-                        })
-                        ->end()
-                        // If there is no connections array key defined, assume a single connection.
-                        ->beforeNormalization()
-                        ->ifTrue(function ($v) {
-                            return \is_array($v) && !\array_key_exists('connections', $v);
-                        })
-                        ->then(function ($v) {
-                            return [
-                                'connections' => [$v],
-                            ];
-                        })
-                        ->end()
                         ->children()
-                            ->arrayNode('connections')
+                            ->arrayNode('hosts')
                                 ->requiresAtLeastOneElement()
-                                ->prototype('array')
-                                    ->fixXmlConfig('header')
-                                    ->children()
-                                        ->scalarNode('url')
-                                            ->validate()
-                                                ->ifTrue(function ($url) {
-                                                    return $url && '/' !== \substr($url, -1);
-                                                })
-                                                ->then(function ($url) {
-                                                    return $url.'/';
-                                                })
-                                            ->end()
-                                        ->end()
-                                        ->scalarNode('username')->end()
-                                        ->scalarNode('password')->end()
-                                        ->scalarNode('host')->end()
-                                        ->scalarNode('port')->end()
-                                        ->scalarNode('proxy')->end()
-                                        ->scalarNode('auth_type')->end()
-                                        ->arrayNode('http_error_codes')
-                                            ->beforeNormalization()
-                                                ->ifTrue(function ($v) { return !\is_array($v); })
-                                                ->then(function ($v) { return [$v]; })
-                                            ->end()
-                                            ->requiresAtLeastOneElement()
-                                            ->defaultValue([400, 403, 404])
-                                            ->prototype('scalar')->end()
-                                        ->end()
-                                        ->scalarNode('aws_access_key_id')->end()
-                                        ->scalarNode('aws_secret_access_key')->end()
-                                        ->scalarNode('aws_region')->end()
-                                        ->scalarNode('aws_session_token')->end()
-                                        ->scalarNode('aws_credential_provider')->end()
-                                        ->booleanNode('ssl')->defaultValue(false)->end()
-                                        ->scalarNode('logger')
-                                            ->defaultValue($this->debug ? 'fos_elastica.logger' : false)
-                                            ->treatNullLike('fos_elastica.logger')
-                                            ->treatTrueLike('fos_elastica.logger')
-                                        ->end()
-                                        ->booleanNode('compression')->defaultValue(false)->end()
-                                        ->arrayNode('headers')
-                                            ->normalizeKeys(false)
-                                            ->useAttributeAsKey('name')
-                                            ->prototype('scalar')->end()
-                                        ->end()
-                                        ->arrayNode('curl')
-                                            ->normalizeKeys(false)
-                                            ->useAttributeAsKey('name')
-                                            ->prototype('scalar')->end()
-                                        ->end()
-                                        ->scalarNode('transport')->end()
-                                        ->scalarNode('timeout')->end()
-                                        ->scalarNode('connectTimeout')->end()
-                                        ->scalarNode('retryOnConflict')
-                                            ->defaultValue(0)
-                                        ->end()
-                                        ->booleanNode('persistent')->defaultValue(true)->end()
+                                ->prototype('scalar')
+                                    ->validate()
+                                    ->ifTrue(function ($url) {
+                                        return $url && !\str_ends_with($url, '/');
+                                    })
+                                    ->then(function ($url) {
+                                        return $url.'/';
+                                    })
                                     ->end()
                                 ->end()
                             ->end()
-                            ->scalarNode('timeout')->end()
-                            ->scalarNode('connectTimeout')->end()
-                            ->scalarNode('headers')->end()
-                            ->scalarNode('connectionStrategy')->defaultValue('Simple')->end()
+                            ->scalarNode('username')->end()
+                            ->scalarNode('password')->end()
+                            ->scalarNode('http_client')->end()
+                            ->scalarNode('cloud_id')->end()
+                            ->scalarNode('retries')->end()
+                            ->scalarNode('api_key')->end()
+                            ->arrayNode('http_error_codes')
+                                ->beforeNormalization()
+                                    ->ifTrue(function ($v) { return !\is_array($v); })
+                                    ->then(function ($v) { return [$v]; })
+                                ->end()
+                                ->requiresAtLeastOneElement()
+                                ->defaultValue([400, 403])
+                                ->prototype('scalar')->end()
+                            ->end()
+                            ->scalarNode('logger')
+                                ->defaultValue($this->debug ? 'fos_elastica.logger' : false)
+                                ->treatNullLike('fos_elastica.logger')
+                                ->treatTrueLike('fos_elastica.logger')
+                            ->end()
+                            ->arrayNode('client_config')
+                                ->children()
+                                    ->scalarNode(RequestOptions::SSL_CERT)->end()
+                                    ->scalarNode(RequestOptions::SSL_KEY)->end()
+                                    ->scalarNode(RequestOptions::SSL_VERIFY)->end()
+                                    ->scalarNode(RequestOptions::SSL_CA)->end()
+                                ->end()
+                            ->end()
+                            ->arrayNode('client_options')
+                                ->normalizeKeys(false)
+                                ->useAttributeAsKey('name')
+                                ->prototype('scalar')->end()
+                            ->end()
+                            ->arrayNode('headers')
+                                ->normalizeKeys(false)
+                                ->useAttributeAsKey('name')
+                                ->prototype('scalar')->end()
+                            ->end()
+                            ->scalarNode('timeout')
+                                ->defaultValue(30)
+                            ->end()
+                            ->scalarNode('retry_on_conflict')
+                                ->defaultValue(0)
+                            ->end()
+                            ->scalarNode('connection_strategy')->defaultValue('Simple')->end()
                         ->end()
                     ->end()
                 ->end()
@@ -519,8 +491,11 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('template_name')
                                 ->info('Defaults to the name of the index template, but can be modified if the index name is different in ElasticSearch')
                             ->end()
-                            ->scalarNode('template')->end()
-                            ->arrayNode('index_patterns')->prototype('scalar')->end()->end()
+                            ->arrayNode('index_patterns')
+                                ->requiresAtLeastOneElement()
+                                ->prototype('scalar')
+                                ->end()
+                            ->end()
                             ->scalarNode('client')->end()
                             ->variableNode('settings')->defaultValue([])->end()
                             ->booleanNode('date_detection')->end()
