@@ -33,7 +33,7 @@ final class ORMPagerProvider implements PagerProviderInterface
         ManagerRegistry $doctrine,
         RegisterListenersService $registerListenersService,
         string $objectClass,
-        array $baseOptions
+        array $baseOptions,
     ) {
         $this->doctrine = $doctrine;
         $this->objectClass = $objectClass;
