@@ -98,9 +98,6 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testObjectsAreTransformedByFindingThemByTheirIdentifiers($elasticaResults, $doctrineObjects)
     {
@@ -120,13 +117,10 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testAnExceptionIsThrownWhenTheNumberOfFoundObjectsIsLessThanTheNumberOfResults(
         $elasticaResults,
-        $doctrineObjects
+        $doctrineObjects,
     ) {
         $transformer = $this->createMockTransformer();
 
@@ -145,13 +139,10 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testAnExceptionIsNotThrownWhenTheNumberOfFoundObjectsIsLessThanTheNumberOfResultsIfOptionSet(
         $elasticaResults,
-        $doctrineObjects
+        $doctrineObjects,
     ) {
         $transformer = $this->createMockTransformer(['ignore_missing' => true]);
 
@@ -169,9 +160,6 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testHighlightsAreSetOnTransformedObjects($elasticaResults, $doctrineObjects)
     {
@@ -194,9 +182,6 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testResultsAreSortedByIdentifier($elasticaResults, $doctrineObjects)
     {
@@ -220,9 +205,6 @@ class AbstractElasticaToModelTransformerTest extends TestCase
 
     /**
      * @dataProvider resultsWithMatchingObjects
-     *
-     * @param mixed $elasticaResults
-     * @param mixed $doctrineObjects
      */
     public function testHybridTransformReturnsDecoratedResults($elasticaResults, $doctrineObjects)
     {
@@ -270,8 +252,6 @@ class AbstractElasticaToModelTransformerTest extends TestCase
     }
 
     /**
-     * @param mixed $options
-     *
      * @return \PHPUnit\Framework\MockObject\MockObject|AbstractElasticaToModelTransformer
      */
     private function createMockTransformer($options = [])
