@@ -11,18 +11,20 @@
 
 namespace FOS\ElasticaBundle\Tests\Unit\Elastica;
 
-use Elastica\Connection;
+use Elastic\Elasticsearch\Exception\ElasticsearchException;
+use Elastic\Elasticsearch\Response\Elasticsearch;
 use Elastica\Exception\ClientException;
 use Elastica\JSON;
 use Elastica\Request;
 use Elastica\Response;
-use Elastica\Transport\NullTransport;
 use FOS\ElasticaBundle\Elastica\Client;
 use FOS\ElasticaBundle\Event\ElasticaRequestExceptionEvent;
 use FOS\ElasticaBundle\Event\PostElasticaRequestEvent;
 use FOS\ElasticaBundle\Event\PreElasticaRequestEvent;
 use FOS\ElasticaBundle\Logger\ElasticaLogger;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Client\ClientInterface;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
@@ -32,7 +34,6 @@ class ClientTest extends TestCase
 {
     public function testRequestsAreLogged()
     {
-        $client = $this->getClientMock();
         $logger = $this->createMock(ElasticaLogger::class);
         $logger
             ->expects($this->once())
@@ -49,16 +50,24 @@ class ClientTest extends TestCase
                 $this->isType('array')
             )
         ;
-        $client->setLogger($logger);
 
-        $response = $client->request('foo');
+        $response = new \GuzzleHttp\Psr7\Response(
+            200,
+            ['Content-Type' => 'application/json', Elasticsearch::HEADER_CHECK => Elasticsearch::PRODUCT_NAME],
+            \json_encode(['foo' => 'bar'], \JSON_THROW_ON_ERROR)
+        );
+        $client = $this->getClient($logger, $response);
 
-        $this->assertInstanceOf(Response::class, $response);
+        $response = $client->sendRequest(new \GuzzleHttp\Psr7\Request(
+            Request::GET,
+            'https://some.tld/foo'
+        ));
+
+        $this->assertInstanceOf(Elasticsearch::class, $response);
     }
 
     public function testSendsNormalEvents(): void
     {
-        $client = $this->getClientMock();
         $dispatcher = $this->createMock(EventDispatcherInterface::class);
         $dispatcher->expects($invoke = $this->exactly(2))
             ->method('dispatch')
@@ -86,11 +95,21 @@ class ClientTest extends TestCase
 
                     $request = $o->getRequest();
 
-                    $this->assertEquals('event', $request->getPath());
-                    $this->assertEquals(Request::GET, $request->getMethod());
-                    $this->assertEquals(['some' => 'data'], $request->getData());
-                    $this->assertEquals(['query' => 'data'], $request->getQuery());
-                    $this->assertEquals(Request::DEFAULT_CONTENT_TYPE, $request->getContentType());
+                    $path = \ltrim($request->getUri()->getPath(), '/'); // to have the same result as in the 6.0
+                    $method = $request->getMethod();
+                    try {
+                        $data = \json_decode((string) $request->getBody(), true, 512, \JSON_THROW_ON_ERROR);
+                    } catch (\JsonException) {
+                        $data = [];
+                    }
+                    $query = [];
+                    \parse_str($request->getUri()->getQuery(), $query);
+
+                    $this->assertEquals('event', $path);
+                    $this->assertEquals(Request::GET, $method);
+                    $this->assertEquals(['some' => 'data'], $data);
+                    $this->assertEquals(['query' => 'data'], $query);
+                    $this->assertEquals(Request::DEFAULT_CONTENT_TYPE, $request->getHeaderLine('Content-Type'));
 
                     $this->assertInstanceOf(Response::class, $o->getResponse());
                 }
@@ -99,28 +118,34 @@ class ClientTest extends TestCase
             }))
         ;
 
+        $response = new \GuzzleHttp\Psr7\Response(
+            200,
+            ['Content-Type' => 'application/json', Elasticsearch::HEADER_CHECK => Elasticsearch::PRODUCT_NAME],
+            \json_encode(['foo' => 'bar'], \JSON_THROW_ON_ERROR)
+        );
+        $logger = $this->createMock(ElasticaLogger::class);
+        $client = $this->getClient($logger, $response);
         $client->setEventDispatcher($dispatcher);
-        $client->request('event', Request::GET, ['some' => 'data'], ['query' => 'data']);
+
+        $client->sendRequest(new \GuzzleHttp\Psr7\Request(
+            Request::GET,
+            'https://some.tld/event?'.\http_build_query(['query' => 'data']),
+            ['Content-Type' => 'application/json'],
+            \json_encode(['some' => 'data'], \JSON_THROW_ON_ERROR)
+        ));
     }
 
     public function testSendsExceptionEvents(): void
     {
         $httpCode = 403;
         $responseString = JSON::stringify(['message' => 'some AWS error']);
-        $transferInfo = [
-            'request_header' => 'bar',
-            'http_code' => $httpCode,
-            'body' => $responseString,
-        ];
-        $response = new Response($responseString);
-        $response->setTransferInfo($transferInfo);
+        $response = new \GuzzleHttp\Psr7\Response(
+            $httpCode,
+            ['Content-Type' => 'application/json', Elasticsearch::HEADER_CHECK => Elasticsearch::PRODUCT_NAME],
+            $responseString
+        );
 
-        $connection = $this->getConnectionMock();
-        $connection->method('getTransportObject')
-            ->willThrowException(new ClientException())
-        ;
-
-        $client = $this->getClientMock($response, $connection);
+        $client = $this->getClient($this->createMock(LoggerInterface::class), $response);
 
         $dispatcher = $this->createMock(EventDispatcherInterface::class);
         $dispatcher->expects($invoke = $this->exactly(2))
@@ -149,13 +174,23 @@ class ClientTest extends TestCase
 
                     $request = $o->getRequest();
 
-                    $this->assertEquals('event', $request->getPath());
-                    $this->assertEquals(Request::GET, $request->getMethod());
-                    $this->assertEquals(['some' => 'data'], $request->getData());
-                    $this->assertEquals(['query' => 'data'], $request->getQuery());
-                    $this->assertEquals(Request::DEFAULT_CONTENT_TYPE, $request->getContentType());
+                    $path = \ltrim($request->getUri()->getPath(), '/'); // to have the same result as in the 6.0
+                    $method = $request->getMethod();
+                    try {
+                        $data = \json_decode((string) $request->getBody(), true, 512, \JSON_THROW_ON_ERROR);
+                    } catch (\JsonException) {
+                        $data = [];
+                    }
+                    $query = [];
+                    \parse_str($request->getUri()->getQuery(), $query);
 
-                    $this->assertInstanceOf(ClientException::class, $o->getException());
+                    $this->assertEquals('event', $path);
+                    $this->assertEquals(Request::GET, $method);
+                    $this->assertEquals(['some' => 'data'], $data);
+                    $this->assertEquals(['query' => 'data'], $query);
+                    $this->assertEquals(Request::DEFAULT_CONTENT_TYPE, $request->getHeaderLine('Content-Type'));
+
+                    $this->assertInstanceOf(ElasticsearchException::class, $o->getException());
                 }
 
                 return true;
@@ -164,40 +199,33 @@ class ClientTest extends TestCase
 
         $client->setEventDispatcher($dispatcher);
         $this->expectException(ClientException::class);
-        $client->request('event', Request::GET, ['some' => 'data'], ['query' => 'data']);
+        $client->sendRequest(new \GuzzleHttp\Psr7\Request(
+            Request::GET,
+            'https://some.tld/event?'.\http_build_query(['query' => 'data']),
+            ['Content-Type' => 'application/json'],
+            \json_encode(['some' => 'data'], \JSON_THROW_ON_ERROR)
+        ));
     }
 
     public function testRequestsWithTransportInfoErrorsRaiseExceptions()
     {
         $httpCode = 403;
         $responseString = JSON::stringify(['message' => 'some AWS error']);
-        $transferInfo = [
-            'request_header' => 'bar',
-            'http_code' => $httpCode,
-            'body' => $responseString,
-        ];
-        $response = new Response($responseString);
-        $response->setTransferInfo($transferInfo);
+        $response = new \GuzzleHttp\Psr7\Response(
+            $httpCode,
+            ['Content-Type' => 'application/json', Elasticsearch::HEADER_CHECK => Elasticsearch::PRODUCT_NAME],
+            $responseString
+        );
 
-        $connection = $this->getConnectionMock();
-        $connection
-            ->expects($this->exactly(1))
-            ->method('hasConfig')
-            ->with('http_error_codes')
-            ->willReturn(true)
-        ;
-        $connection
-            ->expects($this->exactly(1))
-            ->method('getConfig')
-            ->with('http_error_codes')
-            ->willReturn([400, 403, 404])
-        ;
-        $client = $this->getClientMock($response, $connection);
+        $client = $this->getClient($this->createMock(LoggerInterface::class), $response);
 
         $desiredMessage = \sprintf('Error in transportInfo: response code is %d, response body is %s', $httpCode, $responseString);
         $this->expectException(ClientException::class);
         $this->expectExceptionMessage($desiredMessage);
-        $response = $client->request('foo');
+        $client->sendRequest(new \GuzzleHttp\Psr7\Request(
+            Request::GET,
+            'https://some.tld/foo'
+        ));
     }
 
     public function testGetIndexTemplate()
@@ -207,33 +235,14 @@ class ClientTest extends TestCase
         $this->assertSame($template, $client->getIndexTemplate('some_index'));
     }
 
-    private function getConnectionMock()
+    private function getClient(LoggerInterface $logger, \GuzzleHttp\Psr7\Response $response): Client
     {
-        $connection = $this->createMock(Connection::class);
-        $connection->expects($this->any())->method('toArray')->will($this->returnValue([]));
-
-        return $connection;
-    }
-
-    private function getClientMock(?Response $response = null, $connection = null)
-    {
-        $transport = new NullTransport();
-        if ($response) {
-            $transport->setResponse($response);
-        }
-
-        if (!$connection) {
-            $connection = $this->getConnectionMock();
-        }
-        $connection->expects($this->any())->method('getTransportObject')->will($this->returnValue($transport));
-
-        $client = $this->getMockBuilder(Client::class)
-            ->setMethods(['getConnection'])
-            ->getMock()
+        $httpClient = $this->createMock(ClientInterface::class);
+        $httpClient->expects($this->any())
+            ->method('sendRequest')
+            ->willReturn($response)
         ;
 
-        $client->expects($this->any())->method('getConnection')->will($this->returnValue($connection));
-
-        return $client;
+        return new Client(['transport_config' => ['http_client' => $httpClient]], [401, 402, 403], $logger);
     }
 }
