@@ -11,8 +11,8 @@
 
 namespace FOS\ElasticaBundle\Tests\Unit\Index;
 
+use Elastic\Elasticsearch\Endpoints\Indices;
 use Elastica\Client;
-use Elastica\Request;
 use FOS\ElasticaBundle\Configuration\IndexTemplateConfig;
 use FOS\ElasticaBundle\Configuration\ManagerInterface;
 use FOS\ElasticaBundle\Elastica\IndexTemplate;
@@ -41,11 +41,6 @@ class TemplateResetterTest extends TestCase
     private $mappingBuilder;
 
     /**
-     * @var Client&MockObject
-     */
-    private $client;
-
-    /**
      * @var IndexTemplateManager&MockObject
      */
     private $templateManager;
@@ -59,12 +54,10 @@ class TemplateResetterTest extends TestCase
     {
         $this->configManager = $this->createMock(ManagerInterface::class);
         $this->mappingBuilder = $this->createMock(MappingBuilder::class);
-        $this->client = $this->createMock(Client::class);
         $this->templateManager = $this->createMock(IndexTemplateManager::class);
         $this->resetter = new TemplateResetter(
             $this->configManager,
             $this->mappingBuilder,
-            $this->client,
             $this->templateManager
         );
     }
@@ -106,10 +99,9 @@ class TemplateResetterTest extends TestCase
             ->method('create')
             ->with($mapping)
         ;
-        $this->client
+        $indexTemplate
             ->expects($this->never())
-            ->method('request')
-            ->with($this->any(), Request::DELETE)
+            ->method('delete')
         ;
 
         // act
@@ -119,22 +111,28 @@ class TemplateResetterTest extends TestCase
     public function testResetAllIndexesAndDelete()
     {
         // assemble
-        $names = ['first_template'];
+        $templateName = 'index_template';
+        $indexPatterns = ['first_template_index'];
+        $names = [$templateName];
         $mapping = ['properties' => []];
         $this->configManager
             ->method('getIndexNames')
             ->willReturn($names)
         ;
         $indexTemplateConfig = $this->createMock(IndexTemplateConfig::class);
+        $indexTemplateConfig
+            ->method('getName')
+            ->willReturn($templateName)
+        ;
         $this->configManager
             ->method('getIndexConfiguration')
-            ->with('first_template')
+            ->with($templateName)
             ->willReturn($indexTemplateConfig)
         ;
         $indexTemplate = $this->createMock(IndexTemplate::class);
         $this->templateManager
             ->method('getIndexTemplate')
-            ->with('first_template')
+            ->with($templateName)
             ->willReturn($indexTemplate)
         ;
         $this->mappingBuilder
@@ -149,15 +147,29 @@ class TemplateResetterTest extends TestCase
             ->method('create')
             ->with($mapping)
         ;
-        $this->client
+        $indexTemplate
             ->expects($this->once())
-            ->method('request')
-            ->with('first_template/', Request::DELETE)
+            ->method('delete')
+            ->with()
+        ;
+        $indices = $this->createMock(Indices::class);
+        $indices->expects($this->once())
+            ->method('delete')
+            ->with(['index' => $indexPatterns[0].'/'])
+        ;
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('indices')
+            ->willReturn($indices)
+        ;
+        $indexTemplate->expects($this->once())
+            ->method('getClient')
+            ->willReturn($client)
         ;
         $indexTemplateConfig
             ->expects($this->once())
             ->method('getIndexPatterns')
-            ->willReturn(['first_template'])
+            ->willReturn($indexPatterns)
         ;
 
         // act
@@ -192,10 +204,9 @@ class TemplateResetterTest extends TestCase
             ->method('create')
             ->with($mapping)
         ;
-        $this->client
+        $indexTemplate
             ->expects($this->never())
-            ->method('request')
-            ->with($this->any(), Request::DELETE)
+            ->method('delete')
         ;
 
         // act
@@ -205,18 +216,19 @@ class TemplateResetterTest extends TestCase
     public function testResetIndexIndexeAndDelete()
     {
         // assemble
-        $name = 'first_template';
+        $templateName = 'index_template';
+        $indexPatterns = ['first_template_index'];
         $mapping = ['properties' => []];
         $indexTemplateConfig = $this->createMock(IndexTemplateConfig::class);
         $this->configManager
             ->method('getIndexConfiguration')
-            ->with('first_template')
+            ->with($templateName)
             ->willReturn($indexTemplateConfig)
         ;
         $indexTemplate = $this->createMock(IndexTemplate::class);
         $this->templateManager
             ->method('getIndexTemplate')
-            ->with('first_template')
+            ->with($templateName)
             ->willReturn($indexTemplate)
         ;
         $this->mappingBuilder
@@ -231,38 +243,78 @@ class TemplateResetterTest extends TestCase
             ->method('create')
             ->with($mapping)
         ;
-        $this->client
+        $indexTemplate
             ->expects($this->once())
-            ->method('request')
-            ->with('first_template/', Request::DELETE)
+            ->method('delete')
+            ->with()
         ;
         $indexTemplateConfig
             ->expects($this->once())
             ->method('getIndexPatterns')
-            ->willReturn(['first_template'])
+            ->willReturn($indexPatterns)
+        ;
+        $indexTemplateConfig
+            ->method('getName')
+            ->willReturn($templateName)
+        ;
+        $indices = $this->createMock(Indices::class);
+        $indices->expects($this->once())
+            ->method('delete')
+            ->with(['index' => $indexPatterns[0].'/'])
+        ;
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('indices')
+            ->willReturn($indices)
+        ;
+        $indexTemplate->expects($this->once())
+            ->method('getClient')
+            ->willReturn($client)
         ;
 
         // act
-        $this->resetter->resetIndex($name, true);
+        $this->resetter->resetIndex($templateName, true);
     }
 
     public function testDeleteTemplateIndexes()
     {
         // assemble
-        $name = 'some_template';
+        $templateName = 'some_template';
+        $indexPatterns = ['some_template_index'];
         $template = $this->createMock(IndexTemplateConfig::class);
 
         // assert
         $template
             ->expects($this->once())
             ->method('getIndexPatterns')
-            ->willReturn([$name])
+            ->willReturn($indexPatterns)
+        ;
+        $template
+            ->expects($this->once())
+            ->method('getName')
+            ->willReturn($templateName)
         ;
 
-        $this->client
-            ->expects($this->once())
-            ->method('request')
-            ->with('some_template/', Request::DELETE)
+        $indices = $this->createMock(Indices::class);
+        $indices->expects($this->once())
+            ->method('delete')
+            ->with(['index' => $indexPatterns[0].'/'])
+        ;
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('indices')
+            ->willReturn($indices)
+        ;
+
+        $indexTemplate = $this->createMock(IndexTemplate::class);
+        $indexTemplate->expects($this->once())
+            ->method('getClient')
+            ->willReturn($client)
+        ;
+        $this->templateManager
+            ->method('getIndexTemplate')
+            ->with($templateName)
+            ->willReturn($indexTemplate)
         ;
 
         $this->resetter->deleteTemplateIndexes($template);
